api/pyproject.toml
[tool.poetry]
name = "llm-lab-oss"
version = "0.0.1"
description = ""
authors = ["Flowrite team"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
load-dotenv = "^0.1.0"
huggingface-hub = "^0.21.1" #changed from 0.21.4
transformers = "4.38.1"
torch = "^2.2.1"
sse-starlette = "^2.0.0"
flow-merge = { git = "https://github.com/flowritecom/flow-merge.git", rev = "ae8c4108642a9fbce53d2127900c48411d8dd473" }

[tool.poetry.group.dev.dependencies]
debugpy = "^1.6.7.post1"

[tool.poetry.scripts]
serve = "src.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
